feat(rbac): Implement plugin-rh-rbac-common
