+++ plugins/rh-rbac-backend/package.json
@@ -24,31 +24,32 @@
   },
   "dependencies": {
     "@backstage/backend-common": "^0.18.4",
+    "@backstage/catalog-model": "^1.3.0",
     "@backstage/config": "^1.0.7",
     "@backstage/core-plugin-api": "^1.5.2",
+    "@backstage/errors": "^1.2.1",
     "@backstage/plugin-auth-node": "^0.2.14",
     "@backstage/plugin-permission-backend": "^0.5.20",
-    "@backstage/plugin-permission-common": "^0.7.5",
+    "@backstage/plugin-permission-common": "^0.7.7",
     "@backstage/plugin-permission-node": "^0.7.8",
-    "@backstage/catalog-model": "^1.3.0",
-    "@backstage/errors": "^1.2.1",
+    "@janus-idp/plugin-rh-rbac-common": "^0.1.0",
     "@types/express": "*",
     "casbin": "5.26.1",
     "express": "^4.17.1",
     "express-promise-router": "^4.1.0",
     "node-fetch": "^2.6.7",
-    "winston": "^3.2.1",
     "typeorm-adapter": "^1.5.4",
+    "winston": "^3.2.1",
     "yn": "^4.0.0"
   },
   "devDependencies": {
     "@backstage/cli": "^0.22.6",
-    "sinon": "^15.2.0",
     "@types/sinon": "^10.0.15",
     "@types/supertest": "2.0.12",
     "knex": "^2.5.1",
     "knex-mock-client": "^2.0.0",
     "msw": "^1.0.0",
+    "sinon": "^15.2.0",
     "supertest": "^6.2.4"
   },
   "files": [
