+++ plugins/rh-rbac-common-common/README.md
@@ -1,0 +1,5 @@
+# @janus-idp/plugin-rh-rbac-common-common
+
+Welcome to the common package for the rh-rbac-common plugin!
+
+_This plugin was created through the Backstage CLI_
