+++ plugins/rh-rbac-common-common/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@janus-idp/plugin-rh-rbac-common",
+  "description": "Common functionalities for the rh-rbac-common plugin",
+  "version": "0.1.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "private": true,
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "module": "dist/index.esm.js",
+    "types": "dist/index.d.ts"
+  },
+  "backstage": {
+    "role": "common-library"
+  },
+  "scripts": {
+    "build": "backstage-cli package build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test",
+    "clean": "backstage-cli package clean",
+    "prepack": "backstage-cli package prepack",
+    "postpack": "backstage-cli package postpack"
+  },
+  "dependencies": {
+    "@backstage/plugin-permission-common": "^0.7.5"
+  },
+  "devDependencies": {
+    "@backstage/cli": "^0.22.7"
+  },
+  "files": [
+    "dist"
+  ]
+}
