+++ plugins/rh-rbac-common-common/src/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Common functionalities for the rh-rbac plugin.
+ *
+ * @packageDocumentation
+ */
+
+export * from './types';
+export * from './permissions';
