+++ plugins/rh-rbac-common-common/src/permissions.ts
