+++ plugins/rh-rbac-common-common/src/setupTests.ts
@@ -1,0 +1,1 @@
+export {};
