+++ plugins/rh-rbac-common-common/src/types.ts
@@ -1,0 +1,11 @@
+export type PolicyMetadata = {
+  entityReference: string | undefined;
+  permission: string | undefined;
+  policy: string | undefined;
+  effect: string | undefined;
+};
+
+export type UpdatePolicyMetadata = {
+  oldPolicy: PolicyMetadata;
+  newPolicy: PolicyMetadata;
+};
